Drop down to reconnect when status or input change can't be published
In situatuons where broker is outside of network and the connection is closed
the client doesn't understand it needs to reconnect. This makes is to try
to reconnect and also kicks the restart timer. Previously on each loop the
last connect timer was reset as the state was still ON